Creating SearchSchool style file
